test2

+++ print_decimal.c
@@ -13,7 +13,7 @@ int print_decimal(va_list pfList)
 	int sum = 0, x = 1;
 
 	number = va_arg(pfList, int);
-	if (number < 0)
+	if ((int)number < 0)
 	{
 		number *= -1;
 		_putchar('-');
